Added Validation
Added validation so that the user can't go on to the next wizard step without inputing needed data. Added ability to pass validator for each step to the wizard in shipping-label-maker.js. Added logic to steps.js to check the validation when clicking the next button and if their are errors set them to a hook variable. Added validation functions to get-sender-address, get-receiver-address, get-weight, get-service-option, and confirm.